Hacked copy grunt task to include requirejs in ./tmp/pulic/dependencies
This way the sails-linker will automatically add the script to project pages where it is needed. This is not the ideal way to do this, but the linker currently does not allow you to specify other source directories. According to @mikermcneil a new feature is planned for the linker that would make this possible.