Implement BookStore with Backbone
BookStore now functionally inherits from a Backbone Collection and not the EventEmitter. Backbone is an EventEmitter. The Backbone Collection will emit a 'sync' event anytime the Collection is synced with the server. So views can probably, in most cases, update on the sync event, which is marked in BookStore as CHANGE_EVENT.