Make Ajax auth calls in Login and Register forms
I pulled the login & registration AJAX calls to the server that perform the auth actions out of the AuthStore into the React components. This seemed to make sense because the actions in this case do not affect the data layer. They only affect the data layer once the user is logged in. Thus after Login and Register succeed, I then push an action through the dispatcher to the AuthStore to say that the user logged in.